piprequirements.txt
aiofiles==0.3.1
httptools==0.0.9
Jinja2==2.9.5
MarkupSafe==1.0
qrcode==5.3
sanic==0.4.1
six==1.10.0
ujson==1.35
uvloop==0.8.0
